Corrected equation in event 7 of Pre-Publication ZSK Rollover
The equation was: Tret(key) >= Tact(key) + Lzsk.  The ">=" was
replaced by "=".  In this type of rollover, a key becomes retired
as soon as the successor key is made active.  Although the lifetime
of a ZSK is set by policy, Lzsk is the actual lifetime of the key,
the interval between its activation time and the activation time
of the successor.